q2/q3 clipping fixed (hopefully finally except for bsp objects, that is) q2 server should work. added support for addative (and things) blends from menu.dat
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@1035 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/common.c
@@ -707,7 +707,7 @@ coorddata MSG_ToCoord(float f, int bytes)	//return value should be treated as (c
 	switch(bytes)
 	{
 	case 2:
-		r.b2 = LittleShort(f*8);
+		r.b2 = LittleShort((short)(f*8));
 		break;
 	case 4:
 		r.f = LittleFloat(f);
